requirements.txt
PyMuPDF
pytesseract
Pillow
sentence-transformers
faiss-cpu
openai
